pyproject.toml
[project]
name = "numtema-agents-studio"
version = "2.0.0"
description = "Studio d'Agents IA - Plateforme complète pour concevoir, configurer et déployer des agents IA basée sur PocketFlow"
authors = [{ name = "Nümtema", email = "dev@numtema.com" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    # Core Framework
    "pydantic>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    
    # LLM Clients
    "openai>=1.0",
    "anthropic>=0.25",
    "google-genai>=1.9.0",
    
    # Google ADK (optional, for Vertex AI deployment)
    "google-adk>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    
    # MCP Protocol
    "mcp>=1.0",
    
    # Database & Storage
    "sqlalchemy>=2.0",
    "asyncpg>=0.28",
    "redis>=5.0",
    "qdrant-client>=1.7",
    
    # Utils
    "httpx>=0.25",
    "pyyaml>=6.0",
    "python-jose[cryptography]>=3.3",
    "python-dotenv>=1.0.1",
    "passlib[bcrypt]>=1.7",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.5",
    "coverage>=7.0",
]

deployment = [
    "absl-py>=2.2.1",
    "docker>=7.0",
]

[project.optional-dependencies]
all = [
    "langchain>=0.1",
    "chromadb>=0.4",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "RUF"]
ignore = ["E501", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["numtema_studio", "pocketflow", "agents", "tools", "workflows"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
numtema = "numtema_studio.cli:main"
